backend/requirements.txt
fastapi
uvicorn[standard]
langchain-core
langchain-groq
langchain
pydantic
pydantic-settings
pyyaml
faster-whisper
python-multipart
asyncpg
SQLAlchemy[asyncio]
websockets
numpy
pyttsx3
gtts
soundfile
aiofiles